Update script to utilise API v2
I have just discovered there is a way to call the dynamic scenes without using the hue lab alternatives. The way to do this is to use the v2 API over HTTPS. This commit updates the script to always play the dynamic pallette version of a random scene, unless it could not find a dynamic scene; to which it will fallback on the static one. Nonetheless, the script still maintains its stance of being 'dynamic', in the sense that user changes to scenes in the Hue app will be carried over in this script. Only change is it will play dynamic scenes now instead of static ones.